requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.2.6
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
